Refactor from Spectrum into hooks
