+++ packages/@react-aria/color/package.json
@@ -19,11 +19,13 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.6.2",
+    "@react-aria/i18n": "3.1.2",
     "@react-aria/interactions": "^3.2.0",
     "@react-aria/slider": "^3.0.0-alpha.2",
     "@react-aria/utils": "^3.2.1",
     "@react-stately/color": "^3.0.0-alpha.1",
     "@react-types/color": "^3.0.0-alpha.1",
+    "@react-types/shared": "3.2.1",
     "@react-types/slider": "^3.0.0-alpha.1"
   },
   "peerDependencies": {
